Use q-io@2 to exclude wrongly implemented "Array.prototype.find"-method
This method is injected in the `collections`-package as array-shim, which is
a dependency of q-io. Version 1.x of `collections` implements the method with
the semantics of `Array.prototype.findIndex` instead. This leads to problems
with other packages that rely on the implementation as defined in the ES6-Standard

See also: https://github.com/montagejs/collections/issues/139